pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scratchgpt"
version = "0.1.0"
description = "Research-grade from-scratch GPT-style decoder-only Transformer in PyTorch."
readme = "README.md"
requires-python = ">=3.9"
authors = [{ name = "ScratchGPT Authors" }]
dependencies = [
  "torch>=2.0",
  "numpy>=1.23",
  "tqdm>=4.66",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
